Move ProjectionFilters into an internal package
ProjectionFilters is plumbing shared by the blocking and reactor runners
and the annotation registrar, not something callers construct like the
Projection and DcbProjection descriptors. Move it from
org.occurrent.dsl.projection to org.occurrent.dsl.projection.internal so
external users see it is not part of the public API, matching the repo's
internal-package convention. Pure move, no behaviour change.
