+++ dsl/projection-dsl/blocking/src/main/java/org/occurrent/dsl/projection/blocking/Projections.java
@@ -31,7 +31,7 @@ import static java.util.Objects.requireNonNull;
  * Internal helper shared by the blocking projection runners (the Kotlin {@code project} extensions and the Java
  * {@code *ProjectionRunner} facades): turning a {@link Projection} plus a {@link ViewStateRepository} into a
  * {@link MaterializedView} that skips events whose id resolves to {@code null}. The plain-{@code Filter} derivation
- * lives in {@code org.occurrent.dsl.projection.ProjectionFilters}, shared with the reactor stack.
+ * lives in {@code org.occurrent.dsl.projection.internal.ProjectionFilters}, shared with the reactor stack.
  */
 @NullMarked
 public final class Projections {
